project_example/project_example/urls.py
from django.conf.urls import patterns, include, url


urlpatterns = patterns(
    '',
    url(r'^discussions/',
        include('discussions.urls')),
    url(r'^accounts/',
        include('registration.urls')),
)
