project_example/requirements.txt
django-registration
